backend/requirements.txt
asgiref==3.11.1
Django==5.0.14
django-autocomplete-light==3.12.1
django-htmx==1.27.0
django-jazzmin==3.0.2
django-smart-selects==1.7.2
django-suit==0.2.28
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
python-dateutil==2.9.0.post0
